Add files via upload
Working on completing object editors to edit the "Book" Data. Currently Finishing the "Page" Editor to edit pages on a book. This is all being set up for use in the "PageCreator" class (this will be renamed to GUI Creator Class). This class will be the class that uses all the Object oriented Polymorphic classes/objects to easily create and edit a "Book" which is the container for all the sliders/shapes/Text etc.