Add files via upload
Major Improvements on the application. Able to add Quads, Characters, Text -as well as set the text, and add Sliders with the keyboard. Adding Groups on all levels except Page Group is working too. Next Steps will to be working on strengthening the application with more error checking within functions. Also Duplicating and setting position to custom made groupings needs to be completed.